Fix rand function for small ranges.
When the output range is very small `map` will actually result in a
decimal value. Because we're using integers for the calculations the
returned value from `map` is just truncated towards zero. This means that
the only way its possible that the max value of the range is returned,
is when the input value equals to RAND_MAX. Obviously this doesn't
happen very often. To fix this just cast the input to double so the
output integer gets properly rounded.